Refactoring MainScreen component
