+++ project/src/components/main-screen-cities-empty/main-screen-cities-empty.tsx
@@ -1,0 +1,22 @@
+import {useSelector} from 'react-redux';
+import {getCity} from '../../store/user/user-selectors';
+
+function MainScreenCitiesEmpty(): JSX.Element {
+  const currentCity = useSelector(getCity);
+
+  return (
+    <div className="cities">
+      <div className="cities__places-container cities__places-container--empty container">
+        <section className="cities__no-places">
+          <div className="cities__status-wrapper tabs__content">
+            <b className="cities__status">No places to stay available</b>
+            <p className="cities__status-description">We could not find any property available at the moment in {currentCity}</p>
+          </div>
+        </section>
+        <div className="cities__right-section" />
+      </div>
+    </div>
+  );
+}
+
+export default MainScreenCitiesEmpty;
